+++ cmd/project/project_test.go
@@ -1,0 +1,29 @@
+package project_test
+
+import (
+	"os"
+	"testing"
+
+	"github.com/benmatselby/prolificli/cmd/project"
+	"github.com/benmatselby/prolificli/mock_client"
+	"github.com/golang/mock/gomock"
+)
+
+func TestNewWorkspaceCommand(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	defer ctrl.Finish()
+	client := mock_client.NewMockAPI(ctrl)
+
+	cmd := project.NewProjectCommand(client, os.Stdout)
+
+	use := "project"
+	short := "Manage and view your projects in a workspace"
+
+	if cmd.Use != use {
+		t.Fatalf("expected use: %s; got %s", use, cmd.Use)
+	}
+
+	if cmd.Short != short {
+		t.Fatalf("expected use: %s; got %s", short, cmd.Short)
+	}
+}